Read contention between release of DMA and reading of the next DMA descriptor
The original code has a race between the release of DMA buffer ownership and the reading of the next DMA descriptor.
Even if the original code lines were reversed, code generation would still be free to reorder the lines - which could be
the reason why the issue has been latent on gnu++98.  I don't believe that the __DMB() here is actually necessary,
it's the only memory barrier I currently know.  The intention is to preserve the ordering of the next descriptor read and
the release of the buffer.  There's a __DMB() call added after this block in this patch series so the multiple calls are
probably redundant.